Refactor into ScopeView.hs
